+++ vocode/streaming/synthesizer/eleven_labs_synthesizer.py
@@ -49,53 +49,6 @@ class ElevenLabsSynthesizer(BaseSynthesizer[ElevenLabsSynthesizerConfig]):
         self.words_per_minute = 150
         self.experimental_streaming = synthesizer_config.experimental_streaming
 
-    async def experimental_streaming_output_generator(
-        self,
-        response: aiohttp.ClientResponse,
-        chunk_size: int,
-        create_speech_span: Optional[Span],
-    ) -> AsyncGenerator[SynthesisResult.ChunkResult, None]:
-        miniaudio_worker_input_queue: asyncio.Queue[
-            Union[bytes, None]
-        ] = asyncio.Queue()
-        miniaudio_worker_output_queue: asyncio.Queue[
-            Tuple[bytes, bool]
-        ] = asyncio.Queue()
-        miniaudio_worker = MiniaudioWorker(
-            self.synthesizer_config,
-            chunk_size,
-            miniaudio_worker_input_queue,
-            miniaudio_worker_output_queue,
-        )
-        miniaudio_worker.start()
-        stream_reader = response.content
-
-        # Create a task to send the mp3 chunks to the MiniaudioWorker's input queue in a separate loop
-        async def send_chunks():
-            async for chunk in stream_reader.iter_any():
-                miniaudio_worker.consume_nonblocking(chunk)
-            miniaudio_worker.consume_nonblocking(None)  # sentinel
-
-        try:
-            asyncio.create_task(send_chunks())
-
-            # Await the output queue of the MiniaudioWorker and yield the wav chunks in another loop
-            while True:
-                # Get the wav chunk and the flag from the output queue of the MiniaudioWorker
-                wav_chunk, is_last = await miniaudio_worker.output_queue.get()
-                if self.synthesizer_config.should_encode_as_wav:
-                    wav_chunk = encode_as_wav(wav_chunk, self.synthesizer_config)
-
-                yield SynthesisResult.ChunkResult(wav_chunk, is_last)
-                # If this is the last chunk, break the loop
-                if is_last and create_speech_span is not None:
-                    create_speech_span.end()
-                    break
-        except asyncio.CancelledError:
-            pass
-        finally:
-            miniaudio_worker.terminate()
-
     async def create_speech(
         self,
         message: BaseMessage,
@@ -139,7 +92,7 @@ class ElevenLabsSynthesizer(BaseSynthesizer[ElevenLabsSynthesizerConfig]):
             raise Exception(f"ElevenLabs API returned {response.status} status code")
         if self.experimental_streaming:
             return SynthesisResult(
-                self.experimental_streaming_output_generator(
+                self.experimental_mp3_streaming_output_generator(
                     response, chunk_size, create_speech_span
                 ),  # should be wav
                 lambda seconds: self.get_message_cutoff_from_voice_speed(
