+++ vocode/streaming/synthesizer/play_ht_synthesizer.py
@@ -38,6 +38,8 @@ class PlayHtSynthesizer(BaseSynthesizer[PlayHtSynthesizerConfig]):
             raise ValueError(
                 "You must set the PLAY_HT_API_KEY and PLAY_HT_USER_ID environment variables"
             )
+        self.words_per_minute = 150
+        self.experimental_streaming = synthesizer_config.experimental_streaming
 
     async def create_speech(
         self,
@@ -67,11 +69,21 @@ class PlayHtSynthesizer(BaseSynthesizer[PlayHtSynthesizerConfig]):
             f"synthesizer.{SynthesizerType.PLAY_HT.value.split('_', 1)[-1]}.create_total",
         )
 
-        async with self.aiohttp_session.post(
+        response = await self.aiohttp_session.post(
             TTS_ENDPOINT, headers=headers, json=body, timeout=ClientTimeout(total=15)
-        ) as response:
-            if not response.ok:
-                raise Exception(f"Play.ht API error status code {response.status}")
+        )
+        if not response.ok:
+            raise Exception(f"Play.ht API error status code {response.status}")
+        if self.experimental_streaming:
+            return SynthesisResult(
+                self.experimental_mp3_streaming_output_generator(
+                    response, chunk_size, create_speech_span
+                ),  # should be wav
+                lambda seconds: self.get_message_cutoff_from_voice_speed(
+                    message, seconds, self.words_per_minute
+                ),
+            )
+        else:
             read_response = await response.read()
             create_speech_span.end()
             convert_span = tracer.start_span(
